modified: .config/nvim/after/ftplugin/c.lua
new file: .config/nvim/after/ftplugin/cmake.lua

typechange: .config/nvim/after/ftplugin/cpp.lua

new file: .config/nvim/after/ftplugin/dockerfile.lua

modified: .config/nvim/after/ftplugin/python.lua

modified: .config/nvim/after/ftplugin/swift.lua

modified: .config/nvim/lazy-lock.json

modified: .config/nvim/lua/config/project.lua

modified: .config/nvim/lua/config/treesitter.lua

deleted:  .config/nvim/lua/config/venv.lua

modified: .config/nvim/lua/lsp/c.lua

new file: .config/nvim/lua/lsp/cmake.lua

modified: .config/nvim/lua/lsp/python.lua

modified: .config/nvim/lua/lsp/swift.lua

modified: .config/nvim/lua/spec/utilities.lua

modified: .config/yadm/bootstrap.d/python.sh

deleted:  .config/zsh/source/after/ccmake.zsh

new file: .config/zsh/source/after/cmake.zsh

new file: .config/zsh/source/after/swift.zsh

modified: .config/zsh/source/before/plugins.zsh

+++ .config/yadm/bootstrap.d/python.sh
@@ -17,7 +17,7 @@ if ! cmd_exists python3; then
         fi
 fi
 
-if ! cmd_exists pipx; then
+if cmd_exists pip && ! cmd_exists pipx; then
         $PYTHON_EXEC -m pip install --break-system-packages pipx
         pipx ensurepath
 fi
@@ -25,3 +25,5 @@ fi
 if ! cmd_exists virtualenv; then
         pipx install virtualenv
 fi
+
+git clone "https://github.com/MichaelAquilina/zsh-autoswitch-virtualenv.git" "$ZSH_CUSTOM/plugins/autoswitch_virtualenv"
